changed register routes and extracted routes to a different folder
